Creation of CGSolver class, which contains all the methods to solve the given system through the CG method. Code in main is less verbose
