--- conjugate_gradients-main/src/conjugate_gradients.cpp
@@ -1,323 +1,0 @@
-#include <cstdio>
-#include <cstdlib>
-#include <cmath>
-#include <iostream>
-#include <chrono>
-#include <omp.h>
-#include "../include/conjugate_gradients.hpp"
-
-bool read_matrix_from_file(const char *filename, double **matrix_out, size_t *num_rows_out, size_t *num_cols_out)
-{
-    double *matrix;
-    size_t num_rows;
-    size_t num_cols;
-
-    FILE *file = fopen(filename, "rb");
-    if (file == nullptr)
-    {
-        fprintf(stderr, "Cannot open output file\n");
-        return false;
-    }
-
-    fread(&num_rows, sizeof(size_t), 1, file);
-    fread(&num_cols, sizeof(size_t), 1, file);
-    matrix = new double[num_rows * num_cols];
-    fread(matrix, sizeof(double), num_rows * num_cols, file);
-
-    *matrix_out = matrix;
-    *num_rows_out = num_rows;
-    *num_cols_out = num_cols;
-
-    fclose(file);
-
-    return true;
-}
-
-bool write_matrix_to_file(const char *filename, const double *matrix, size_t num_rows, size_t num_cols)
-{
-    FILE *file = fopen(filename, "wb");
-    if (file == nullptr)
-    {
-        fprintf(stderr, "Cannot open output file\n");
-        return false;
-    }
-
-    fwrite(&num_rows, sizeof(size_t), 1, file);
-    fwrite(&num_cols, sizeof(size_t), 1, file);
-    fwrite(matrix, sizeof(double), num_rows * num_cols, file);
-
-    fclose(file);
-
-    return true;
-}
-
-void print_matrix(const double *matrix, size_t num_rows, size_t num_cols, FILE *file = stdout)
-{
-    fprintf(file, "%zu %zu\n", num_rows, num_cols);
-    for (size_t r = 0; r < num_rows; r++)
-    {
-        for (size_t c = 0; c < num_cols; c++)
-        {
-            double val = matrix[r * num_cols + c];
-            printf("%+6.3f ", val);
-        }
-        printf("\n");
-    }
-}
-
-// dot product between two vectors
-double dot(const double *x, const double *y, size_t size)
-{
-    double result = 0.0;
-// also take into account omp simd directive
-#pragma omp parallel for simd reduction(+ : result)
-    for (size_t i = 0; i < size; i++)
-    {
-        result += x[i] * y[i];
-    }
-    return result;
-}
-
-#pragma omp declare simd
-void axpby(double alpha, const double *x, double beta, double *y, size_t size)
-{
-    // y = alpha * x + beta * y
-    // #pragma omp target map(to : size) map(tofrom : x)
-    // #pragma omp teams num_teams(4) thread_limit(size / 2)
-    // #pragma omp distribute parallel for simd
-#pragma omp parallel for simd
-    for (size_t i = 0; i < size; i++)
-    {
-        y[i] = alpha * x[i] + beta * y[i];
-    }
-    // #pragma omp end distribute parallel for simd
-    // #pragma omp end teams
-    // #pragma omp end target
-}
-
-// method to compute the new direction, which ensures global A-orthogonality property
-// void gemv(double alpha, const double *A, const double *x, double beta, double *y, size_t num_rows, size_t num_cols)
-// {
-//     // y = alpha * A * x + beta * y;
-
-//     // pragma omp parallel for collapse(2)
-//     for (size_t r = 0; r < num_rows; r++)
-//     {
-//         double y_val = 0.0;
-//         for (size_t c = 0; c < num_cols; c++)
-//         {
-//             y_val += alpha * A[r * num_cols + c] * x[c];
-//         }
-//         y[r] = beta * y[r] + y_val;
-//     }
-// }
-
-void precA(const double *A, const double *x, double *Ax, size_t size)
-{
-#pragma omp parallel for
-    for (size_t i = 0; i < size; i++)
-    {
-        double y_val = 0.0;
-        for (size_t j = 0; j < size; j++)
-        {
-            y_val += A[i * size + j] * x[j];
-        }
-        Ax[i] = y_val;
-    }
-}
-
-void conjugate_gradients(const double *A, const double *b, double *x, size_t size, int max_iters, double rel_error)
-{
-    using namespace std::chrono;
-
-    double alpha, beta, bb, rr, rr_new;
-    // residual
-    double *r = new double[size];
-    // preconditioned residual
-    double *p = new double[size];
-    double *Ap = new double[size];
-    int num_iters;
-
-    // Get starting timepoint
-    auto start = high_resolution_clock::now();
-
-#pragma omp parallel for simd
-    for (size_t i = 0; i < size; i++)
-    {
-        x[i] = 0.0;
-        r[i] = b[i];
-        p[i] = b[i];
-    }
-
-    // needed for stopping criterion
-    bb = dot(b, b, size);
-    rr = bb;
-
-    for (num_iters = 1; num_iters <= max_iters; num_iters++)
-    {
-        // brainy way to compute A * p, need it for residual update and computation of alpha
-        // writes result directly in Ap
-        // is this some kind of obfuscation to make the code less readable???
-        // gemv(1.0, A, p, 0.0, Ap, size, size);
-        precA(A, p, Ap, size);
-        // compute new alpha coefficient to guarantee optimal convergence rate
-        alpha = rr / dot(p, Ap, size);
-        
-#pragma omp parallel sections
-        {
-#pragma omp section
-            {
-                // compute new approximate of the solution at step k+1
-                // x_k+1 = x_k + alpha_k * p_k
-                axpby(alpha, p, 1.0, x, size);
-            }
-#pragma omp section
-            {
-                // compute new residual at step k+1
-                // r_k+1 = r_k - alpha_k * A * p_k
-                axpby(-alpha, Ap, 1.0, r, size);
-            }
-        }
-
-        // update the 2-norm of the residual at step k+1
-        rr_new = dot(r, r, size);
-        // compute beta coefficient
-        // beta_k = ||r_k+1||^2 / ||r_k||^2
-        // stopping criterion ==> sqrt(||r||^2 / ||b||^2) < rel_error equivalent to 2-norm or euclidean norm
-        beta = rr_new / rr;
-        rr = rr_new;
-        if (std::sqrt(rr / bb) < rel_error)
-        {
-            break;
-        }
-
-        // compute new direction at step k+1
-        // p_k+1 = r_k+1 + beta_k * p_k
-        axpby(1.0, r, beta, p, size);
-    }
-
-    auto stop = high_resolution_clock::now();
-    // // Get duration. Substart timepoints to
-    // // get duration. To cast it to proper unit
-    // // use duration cast method
-    auto duration = duration_cast<milliseconds>(stop - start);
-    std::cout << "Time taken by function: "
-              << duration.count() << " milliseconds" << std::endl;
-
-    delete[] r;
-    delete[] p;
-    delete[] Ap;
-
-    if (num_iters <= max_iters)
-    {
-        printf("Converged in %d iterations, relative error is %e\n", num_iters, std::sqrt(rr / bb));
-    }
-    else
-    {
-        printf("Did not converge in %d iterations, relative error is %e\n", max_iters, std::sqrt(rr / bb));
-    }
-}
-
-int main(int argc, char **argv)
-{
-    printf("Usage: ./random_matrix input_file_matrix.bin input_file_rhs.bin output_file_sol.bin max_iters rel_error\n");
-    printf("All parameters are optional and have default values\n");
-    printf("\n");
-
-    const char *input_file_matrix = "io/matrix.bin";
-    const char *input_file_rhs = "io/rhs.bin";
-    const char *output_file_sol = "io/sol.bin";
-    int max_iters = 1000;
-    double rel_error = 1e-9;
-
-    if (argc > 1)
-        input_file_matrix = argv[1];
-    if (argc > 2)
-        input_file_rhs = argv[2];
-    if (argc > 3)
-        output_file_sol = argv[3];
-    if (argc > 4)
-        max_iters = atoi(argv[4]);
-    if (argc > 5)
-        rel_error = atof(argv[5]);
-
-    printf("Command line arguments:\n");
-    printf("  input_file_matrix: %s\n", input_file_matrix);
-    printf("  input_file_rhs:    %s\n", input_file_rhs);
-    printf("  output_file_sol:   %s\n", output_file_sol);
-    printf("  max_iters:         %d\n", max_iters);
-    printf("  rel_error:         %e\n", rel_error);
-    printf("\n");
-
-    double *matrix;
-    double *rhs;
-    size_t size;
-
-    {
-        printf("Reading matrix from file ...\n");
-        size_t matrix_rows;
-        size_t matrix_cols;
-        bool success_read_matrix = read_matrix_from_file(input_file_matrix, &matrix, &matrix_rows, &matrix_cols);
-        if (!success_read_matrix)
-        {
-            fprintf(stderr, "Failed to read matrix\n");
-            return 1;
-        }
-        printf("Done\n");
-        printf("\n");
-
-        printf("Reading right hand side from file ...\n");
-        size_t rhs_rows;
-        size_t rhs_cols;
-        bool success_read_rhs = read_matrix_from_file(input_file_rhs, &rhs, &rhs_rows, &rhs_cols);
-        if (!success_read_rhs)
-        {
-            fprintf(stderr, "Failed to read right hand side\n");
-            return 2;
-        }
-        printf("Done\n");
-        printf("\n");
-
-        if (matrix_rows != matrix_cols)
-        {
-            fprintf(stderr, "Matrix has to be square\n");
-            return 3;
-        }
-        if (rhs_rows != matrix_rows)
-        {
-            fprintf(stderr, "Size of right hand side does not match the matrix\n");
-            return 4;
-        }
-        if (rhs_cols != 1)
-        {
-            fprintf(stderr, "Right hand side has to have just a single column\n");
-            return 5;
-        }
-
-        size = matrix_rows;
-    }
-
-    printf("Solving the system ...\n");
-    double *sol = new double[size];
-    conjugate_gradients(matrix, rhs, sol, size, max_iters, rel_error);
-    printf("Done\n");
-    printf("\n");
-
-    printf("Writing solution to file ...\n");
-    bool success_write_sol = write_matrix_to_file(output_file_sol, sol, size, 1);
-    if (!success_write_sol)
-    {
-        fprintf(stderr, "Failed to save solution\n");
-        return 6;
-    }
-    printf("Done\n");
-    printf("\n");
-
-    delete[] matrix;
-    delete[] rhs;
-    delete[] sol;
-
-    printf("Finished successfully\n");
-
-    return 0;
-}
